Split utils into submods
